+++ src/component/layoutComponent/header/Header.jsx
@@ -1,11 +1,16 @@
 import React from "react";
-import { useMobileLayout } from "../../customHooks/index";
+import { useMobileLayout } from "../../../customHooks/index";
 
 const Header = () => {
   var a = "hello";
   const displayMobileLayout = useMobileLayout();
   console.log(displayMobileLayout, a);
-  return <div>Header</div>;
+  if (displayMobileLayout) return <h1>mobile layout</h1>;
+  return (
+    <header>
+      <nav></nav>
+    </header>
+  );
 };
 
 export default Header;
